[REF] ps: split base.py
