+++ product_subscription/models/__init__.py
@@ -1,7 +1,8 @@
-from . import base
+from . import account_move_line
 from . import invoice
+from . import product_release
+from . import product_template
+from . import res_partner
 from . import subscription_template
 from . import subscription_object
 from . import subscription_request
-from . import product_release
-from . import account_move_line
